src/binance-api-update-lambda/requirements.txt
python-binance 
python-dotenv
Requests
boto3
pandas
logging
psycopg2-binary